Use float64 when solving the optimization

+++ tensor_splines/integral.py
@@ -85,7 +85,9 @@ def _integrate_spiro_12th_degree_n_parts(k_parameters: torch.Tensor, n: int):
         W = torch.tensor([[1., s, .5 * s2, (1./6) * s3],
                           [0., ds, ds * s, .5 * ds * s2],
                           [0., 0., ds2, s * ds2],
-                          [0., 0., 0., ds3]], device=k_parameters.device)
+                          [0., 0., 0., ds3]],
+                         dtype=torch.float64,
+                         device=k_parameters.device)
         km = torch.einsum('...ij,...j->...i', W, k)
         u, v = _integrate_spiro_12th_degree(km)
